# make calling GapObj function objects faster when no keyword arguments are involved