--- a/src/ccalls.jl
+++ b/src/ccalls.jl
@@ -187,16 +187,25 @@ function call_gap_func(func::GapObj, args...; kwargs...)
     end
 end
 
+# specialize call_gap_func for the no-keywords case, for performance
+function call_gap_func_nokw(func::GapObj, args...)
+    if TNUM_OBJ(func) == T_FUNCTION && length(args) <= 6
+        _call_gap_func(func, args...)
+    else
+        ccall(:call_gap_func, Any, (Any, Any), func, args)
+    end
+end
+
 # make all GapObj callable
 (func::GapObj)(args...; kwargs...) = call_gap_func(func, args...; kwargs...)
 
 # specialize non-kwargs versions, which increases performance
-(func::GapObj)(a1) = call_gap_func(func, a1)
-(func::GapObj)(a1, a2) = call_gap_func(func, a1, a2)
-(func::GapObj)(a1, a2, a3) = call_gap_func(func, a1, a2, a3)
-(func::GapObj)(a1, a2, a3, a4) = call_gap_func(func, a1, a2, a3, a4)
-(func::GapObj)(a1, a2, a3, a4, a5) = call_gap_func(func, a1, a2, a3, a4, a5)
-(func::GapObj)(a1, a2, a3, a4, a5, a6) = call_gap_func(func, a1, a2, a3, a4, a5, a6)
+(func::GapObj)(a1) = call_gap_func_nokw(func, a1)
+(func::GapObj)(a1, a2) = call_gap_func_nokw(func, a1, a2)
+(func::GapObj)(a1, a2, a3) = call_gap_func_nokw(func, a1, a2, a3)
+(func::GapObj)(a1, a2, a3, a4) = call_gap_func_nokw(func, a1, a2, a3, a4)
+(func::GapObj)(a1, a2, a3, a4, a5) = call_gap_func_nokw(func, a1, a2, a3, a4, a5)
+(func::GapObj)(a1, a2, a3, a4, a5, a6) = call_gap_func_nokw(func, a1, a2, a3, a4, a5, a6)
 
 #
 # below several "fastpath" methods for call_gap_func follow which directly