@fingolfin 2020-10-14
First off: thanks for this.

Second, that's a *major* regression compared to Julia 1.5 then! Ouch! Your PR partially repairs that, but not quite. I'll post my numbers in a separate comment.

Third: One major issue with GAP.jl is that we use "foreign types" in Julia, and those do not play well with precompilation etc. Since the concrete type (MPtr) is only known during runtime, a lot of things in GAP.jl are less efficient than they otherwise could be. One (perhaps far-fetched) idea here is that perhaps one could extend Julia a bit to allow something like `foreign type MPtr <: GapObj end` to "predeclare" a foreign type, so that the precompiler deal with it, even though it is only actually defined at runtime. I don't know what the precompiler really needs, though (I imagine at the very least the supertype and size of an instance -- both we could provide. If that would allow me to write the concrete type `MPtr` instead of the abstract type `GapObj` in a bunch of cases, that would already help.

Not having that, I thought about various ways to "trick" Julia into still allowing me to handle things efficiently, e.g. using parametric types. But in practice, all my experiments became quickly rather unwildly, and/or were confusing to use from outside the package and/or still required outside code using GAP.jl to employ similar tricks.

So I gave up on that for the time being (there are too many other things I need to take of). But I'd be interested in sitting down with you, @rfourquet, to discuss this -- perhaps we can come up with a way to do things better. Also would be happy to discuss this with other Julia devs, but I didn't really know where to start... Oh, see also https://github.com/oscar-system/Oscar.jl/issues/128

Another potential way to gain performance is the following: there are some functions in the GAP kernel which map strings to unique integers. That is, the first time you pass in a string, a unique integer is assigned to it, and stored in a hashtable. Any subsequent invocation with an equal string just performs a hashtable lookup and returns the same integer... This is used behind the scenes to implement the `getproperty` etc. methods for `GAP.Globals`. Now, each time you use `GAP.Globals.FOO`, it has to call into the GAP kernel to map `FOO` to this integer. This is very wasteful. What I'd like to happen is that the first time Julia encounters `GAP.Globals.FOO`, it performs that lookup, causing the GAP kernel to assign that unique integer; and then any subsequent occurrence of `GAP.Globals.FOO` could be transformed into using that now-constant integer value corresponding to `FOO` directly. But the only ways to achieve that that I could think of involve (re-)evaluating any methods using `GAP.Globals.FOO` at runtime... Or perhaps to actually only evaluate such methods at runtime, not (pre)compile time.

For a time, I thought I could (ab)use `@pure` to do this, but that doesn't really work (see also 04e8b2eade7925a8da19928f2efa817f0f7bac93 and 0472308d7e0af3ed8219ce6e067e84f97044fb1c for the commits that revert the short lived code).

Anyway, as a consequence, doing things like `IsFinite = GAP.Globals.IsFinite` and using that instead can help a lot. My benchmark numbers in the next comment will hence include those.