@fingolfin 2020-10-14
Here is what I get on my machine with Julia 1.5.2:
```
julia> using GAP, BenchmarkTools

julia> IsFinite = GAP.Globals.IsFinite;

julia> l = GAP.evalstr( "[1,2,3]" );

julia> @btime GAP.Globals.IsFinite($l);
  141.825 ns (1 allocation: 16 bytes)

julia> @btime IsFinite($l);
  104.670 ns (1 allocation: 16 bytes)

julia> @btime $l[1];
  696.676 ns (8 allocations: 192 bytes)

julia> @eval GAP function call_gap_func(func::GapObj, args...; kwargs...) ...

julia> @btime GAP.Globals.IsFinite($l);
  142.923 ns (1 allocation: 16 bytes)

julia> @btime IsFinite($l);
  104.374 ns (1 allocation: 16 bytes)

julia> @btime $l[1];
  732.634 ns (8 allocations: 192 bytes)
```

As you can see, re-evaluating `call_gap_func` makes no difference here.

And here is what I get with Julia master as of this morning. Observe that calls to `IsFinite` got substantially slower, while `l[1]` got faster:
```
julia> using GAP, BenchmarkTools

julia> IsFinite = GAP.Globals.IsFinite;

julia> l = GAP.evalstr( "[1,2,3]" );

julia> @btime GAP.Globals.IsFinite($l);
  592.313 ns (8 allocations: 128 bytes)

julia> @btime IsFinite($l);
  554.021 ns (8 allocations: 128 bytes)

julia> @btime $l[1];
  632.236 ns (8 allocations: 192 bytes)

julia> @eval GAP function call_gap_func(func::GapObj, args...; kwargs...) ...

julia> @btime GAP.Globals.IsFinite($l);
  397.085 ns (5 allocations: 80 bytes)

julia> @btime IsFinite($l);
  354.870 ns (5 allocations: 80 bytes)

julia> @btime $l[1];
  640.646 ns (8 allocations: 192 bytes)
```