@fingolfin 2020-10-14
With your changes (which I inserted into my active Julia sessions using `@eval GAP ...`, I get these numbers in Julia 1.5.2:
```
julia> @btime GAP.Globals.IsFinite($l);
  76.358 ns (0 allocations: 0 bytes)

julia> @btime IsFinite($l);
  44.101 ns (0 allocations: 0 bytes)

julia> @btime $l[1];
  62.442 ns (0 allocations: 0 bytes)
```
and these in Julia master:
```
julia> @btime GAP.Globals.IsFinite($l);
  69.618 ns (0 allocations: 0 bytes)

julia> @btime IsFinite($l);
  44.859 ns (0 allocations: 0 bytes)

julia> @btime $l[1];
  54.712 ns (0 allocations: 0 bytes)
```

So, this helps a lot on both Julia versions, and also fixes the sever performance regression with Julia master. And `l[1]` got even faster on Julia master.

For the function calls it might look as if it got faster, too, but I did these measurements on my laptop and fluctuations in the benchmarks are quite high, so I'd be careful about drawing conclusions; I ought to redo these on a benchmark server.  But in the meantime, `@benchmark` of course gives some more info, and that looks promising... with Julia 1.5.2:
```
julia> @benchmark IsFinite($l)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     46.879 ns (0.00% GC)
  median time:      48.424 ns (0.00% GC)
  mean time:        50.552 ns (0.00% GC)
  maximum time:     239.901 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     984
```
And with Julia master:
```
julia> @benchmark IsFinite($l)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     44.773 ns (0.00% GC)
  median time:      46.164 ns (0.00% GC)
  mean time:        47.776 ns (0.00% GC)
  maximum time:     180.966 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     985
```