@fingolfin 2020-10-14
BTW here is another optimization I'd like to have: in the above example, I set `IsFinite = GAP.Globals.IsFinite`. From that point on `IsFinite` represent a `T_FUNCTION`, i.e. I *know* that we will end up calling `_call_gap_func`; this implies another tiny performance gain is possible:
```
julia> @btime IsFinite($l);
  41.270 ns (0 allocations: 0 bytes)

julia> @btime GAP._call_gap_func(IsFinite, $l);
  38.013 ns (0 allocations: 0 bytes)
```

Of course one natural way to get *that* speedup would be to use a distinct type for `T_FUNCTION` etc.; right now in GAP, all GAP objects have the same type `MPtr`, but in theory there could be multiple types for special GAP "types", such as `T_FUNCTION` etc. (but one has to be *very* careful there, as not for all GAP "types" is this information truly constant... but it is for `T_FUNCTION`). Alas, changing that is a non-trivial endeavour.